moved things around to be more standard
